apps/backend/pyproject.toml
[project]
name = "bdns-api"
version = "0.1.0"
dependencies = [
    "fastapi",
    "uvicorn",
    "strawberry-graphql",
    "sqlalchemy",
    "alembic",
    "bdns_core",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
